Script/requirements.txt
colored==1.4.3
dnspython==2.2.1
prettytable==2.1.0
yaspin==2.2.0
